# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Bot Atendimento Life API
  version: 0.1.0
  description: |
    Contrato da API do orquestrador de triagem WhatsApp da Life Engenharia.
    Endpoints marcados como `planned` documentam o contrato da próxima entrega e não estão disponíveis ainda.
    A interface Swagger é publicada localmente em `/documentation`.
  contact:
    name: Time de Desenvolvimento Life
servers:
  - url: http://localhost:3000
    description: Desenvolvimento local
...
paths:
...
  /webhooks/whatsapp:
    get:
      tags: [WhatsApp webhook]
      summary: Valida a assinatura do webhook da Meta
      description: A Meta chama este endpoint durante o cadastro do callback.
      parameters:
        - $ref: '#/components/parameters/HubMode'
        - $ref: '#/components/parameters/HubVerifyToken'
        - $ref: '#/components/parameters/HubChallenge'
      responses:
        '200':
          description: Desafio retornado em texto puro
          content:
            text/plain:
              schema: { type: string }
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [WhatsApp webhook]
      summary: Recebe eventos normalizados do WhatsApp
      description: |
        Implementado atualmente para o contrato de simulação. O adaptador Meta converterá o payload oficial
        para este formato interno antes de processar a conversa. A mensagem é idempotente pelo campo `id`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookSimulationRequest'
      responses:
        '200':
          description: Evento aceito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookReceipt'
...
components:
  parameters:
    HubMode:
      name: hub.mode
      in: query
      required: true
      schema: { type: string, example: subscribe }
    HubVerifyToken:
      name: hub.verify_token
      in: query
      required: true
      schema: { type: string }
    HubChallenge:
      name: hub.challenge
      in: query
      required: true
      schema: { type: string }
...
  responses:
    Forbidden:
      description: Token de verificação inválido
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    IncomingMessage:
      type: object
      required: [id, from]
      properties:
        id: { type: string, example: wamid.HBgMNTUxMTk5OTk5OTk5OQ== }
        from: { type: string, example: '5511999999999', description: Telefone em formato E.164 sem sinal de mais. }
        text: { type: string, example: Olá }
    WebhookSimulationRequest:
      type: object
      properties:
        messages:
          type: array
          items: { $ref: '#/components/schemas/IncomingMessage' }
    WebhookReceipt:
      type: object
      required: [received, processed, duplicates]
      properties:
        received: { type: boolean, example: true }
        processed: { type: integer, example: 1 }
        duplicates: { type: integer, example: 0 }
        responses:
          type: array
          items: { $ref: '#/components/schemas/MenuResponse' }
    MenuResponse:
      type: object
      required: [to, state, text]
      properties:
        to: { type: string, example: '5511999999999' }
        state: { type: string, enum: [CONSENT, ROUTE, COMMERCIAL_SERVICE_MENU, COMMERCIAL_COLLECTING, TECHNICAL_COLLECTING, HANDOFF_HUMANO] }
        text: { type: string }
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...